+++ idb_push.py
@@ -8,9 +8,10 @@ import zmq_primitives
 import hooks
 import idb_push_ui
 import idb_push_ops
-import idb_push_config
-reload(idb_push_config)
+
 from idb_push_config import *
+CONFIGURATION = get_configuration()
+
 # Reload hack for debug sessions
 if CONFIGURATION[DEBUG]:
     reload(psida_common)
@@ -19,9 +20,9 @@ if CONFIGURATION[DEBUG]:
     reload(idb_push_ui)
     reload(idb_push_ops)
 
-# Globals for holding IDB and IDP hook currently registered
-g_idp_hook = hooks.IDPHook()
+# Global for holding IDB hooks currently registered
 g_idb_hook = hooks.IDBHook()
+g_hooks_enabled = hooks.g_hooks_enabled
 
 # Holds the current thread running for receiving incoming updates
 g_receive_thread = None
@@ -83,9 +84,6 @@ def start():
     # open global socket
     hooks.g_zmq_socket = zmq_primitives.zmq_open_pub_socket()  # default arguments
 
-    if not g_idp_hook.hook():
-        raise Exception('IDPHook installation FAILED')
-
     if not g_idb_hook.hook():
         raise Exception('IDBHook installation FAILED')
 
@@ -107,7 +105,6 @@ def start():
 
 def _remove_hooks_and_stop_thread():
     hooks.g_hooks_enabled = False
-    g_idp_hook.unhook()
     g_idb_hook.unhook()
 
     global g_receive_thread
@@ -139,5 +136,5 @@ def stop(reason=None):
 
     global g_form
     if g_form is not None and reason != idaapi.NW_TERMIDA:
-        g_form.Close(idaapi.PluginForm.FORM_SAVE)
+        g_form.Close(idaapi.PluginForm.WOPN_RESTORE)